+++ src/backend/outbuf.c
@@ -143,11 +143,11 @@ void Outbuffer::writeByte(int v)
 /**
  * Writes a 32 bit int.
  */
-void Outbuffer::write32(long v)
+void Outbuffer::write32(int v)
 {
     if (pend - p < 4)
         reserve(4);
-    *(long *)p = v;
+    *(int *)p = v;
     p += 4;
 }
 
@@ -260,7 +260,7 @@ void Outbuffer::setsize(unsigned size)
     p = buf + size;
 }
 
-void Outbuffer::writesLEB128(long value)
+void Outbuffer::writesLEB128(int value)
 {
     int negative = (value < 0);
 
@@ -279,7 +279,7 @@ void Outbuffer::writesLEB128(long value)
     }
 }
 
-void Outbuffer::writeuLEB128(unsigned long value)
+void Outbuffer::writeuLEB128(unsigned value)
 {
     do
     {   unsigned char b = value & 0x7F;
